quiz/requirements.txt
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
openai==1.43.0
python-dotenv==1.0.1
pydantic==2.9.2
rich==13.8.1
httpx==0.27.2
